Исправить ошибки в контроллере "Сотрудников". Реализовать 1 форму и контроллер для обновления и добавления.
