+++ src/main/resources/db/filedb.trace.db
@@ -1036,7 +1036,7 @@ Database is already closed (to disable automatic closing at VM shutdown, add ";D
 	at org.hibernate.cfg.Configuration.buildSessionFactory(Configuration.java:728)
 	at org.hibernate.cfg.Configuration.buildSessionFactory(Configuration.java:746)
 	at com.budget.repository.EmployeeRepository.<init>(EmployeeRepository.java:20)
-	at com.budget.controllers.service.UpdateEmployeeController.<init>(UpdateEmployeeController.java:19)
+	at com.budget.controllers.service.EmployeeController.<init>(UpdateEmployeeController.java:19)
 	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
 	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
@@ -1156,7 +1156,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:C:/User
 	at org.hibernate.cfg.Configuration.buildSessionFactory(Configuration.java:728)
 	at org.hibernate.cfg.Configuration.buildSessionFactory(Configuration.java:746)
 	at com.budget.repository.EmployeeRepository.<init>(EmployeeRepository.java:20)
-	at com.budget.controllers.service.UpdateEmployeeController.<init>(UpdateEmployeeController.java:19)
+	at com.budget.controllers.service.EmployeeController.<init>(UpdateEmployeeController.java:19)
 	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
 	at java.base/jdk.internal.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
@@ -1291,7 +1291,7 @@ Database is already closed (to disable automatic closing at VM shutdown, add ";D
 	at org.hibernate.internal.SessionImpl.saveOrUpdate(SessionImpl.java:645)
 	at org.hibernate.internal.SessionImpl.saveOrUpdate(SessionImpl.java:640)
 	at com.budget.repository.EmployeeRepository.addOrUpdateEmployee(EmployeeRepository.java:25)
-	at com.budget.controllers.service.UpdateEmployeeController.updateEmployee(UpdateEmployeeController.java:99)
+	at com.budget.controllers.service.EmployeeController.updateEmployee(UpdateEmployeeController.java:99)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
@@ -1409,7 +1409,7 @@ General error: "java.lang.IllegalStateException: The file is locked: nio:C:/User
 	at org.hibernate.internal.SessionImpl.saveOrUpdate(SessionImpl.java:645)
 	at org.hibernate.internal.SessionImpl.saveOrUpdate(SessionImpl.java:640)
 	at com.budget.repository.EmployeeRepository.addOrUpdateEmployee(EmployeeRepository.java:25)
-	at com.budget.controllers.service.UpdateEmployeeController.updateEmployee(UpdateEmployeeController.java:99)
+	at com.budget.controllers.service.EmployeeController.updateEmployee(UpdateEmployeeController.java:99)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
 	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
 	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
