+++ src/main/resources/view/employee/dashboardEmployee.fxml
@@ -36,7 +36,7 @@
                         <Font name="Impact" size="16.0"/>
                     </font>
                 </Button>
-                <Button id="updateButton" alignment="CENTER" mnemonicParsing="false" onAction="#updateItem"
+                <Button id="updateButton" alignment="CENTER" mnemonicParsing="false" onAction="#updateEmployee"
                         prefHeight="55.0" prefWidth="132.0" text="Изменить">
                     <VBox.margin>
                         <Insets bottom="5.0" left="5.0" right="5.0" top="5.0"/>
