+++ src/main/resources/view/employee/updateEmployee.fxml
@@ -8,7 +8,7 @@
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
             nodeOrientation="LEFT_TO_RIGHT" prefHeight="400.0" prefWidth="600.0"
             xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.budget.controllers.service.UpdateEmployeeController">
+            fx:controller="com.budget.controllers.service.EmployeeController">
     <VBox prefHeight="400.0" prefWidth="600.0">
         <HBox prefHeight="58.0" prefWidth="600.0">
             <Label alignment="CENTER" prefHeight="67.0" prefWidth="600.0" text="Изменить данные сотрудника"
